Add new workflow for manual QA suite runs
This commit adds a new Github workflow which can be triggered manually
(through the 'workflow_dispatch' event trigger). It allows to run the QA
suite against a given branch (or a fork of Ganeti with a given branch).
It can be triggered via the 'Run workflow' button on the 'Actions' page
and will prompt for all relevant inputs.

Signed-off-by: Rudolph Bott <r@spam.wtf>