move-instance: allow users to empty out NIC parameters
In #1696, I ended up in a situation where the instance I'm moving had
both "network" and "link" parameters. That seems like a natural
configuration in our cluster, but something move-instance really
doesn't like, as it's passing this verbatim to create-instance which
naturally freaks out on the contradictory arguments.

Instead of just crashing, detect empty parameters (`not v` is a
shortcut here) and null them out. This will result in those parameters
being passed as empty to create-instance, or to be more accurate, not
be passed at all.

I've been able to fix the crash in #1696 by passing `--net
0:ip=pool,network=gnt-dal-01,mode=,link=` to move-instance with this
patch.

I'm not sure it's the right approach though. I'd much rather *not*
have to pass `--net` at all, as I actually want to move multiple
instances across clusters, and that seems incompatible with `--net`
for some reason I cannot currently fathom (but which is possibly
related to this problem).