Qemu: handle additional parameters appended to the `-name` argument
While trying to introduce the `debug_threads` option (which gets appended
to the machine name) we broke clusters after a downgrade (see #1820). We
removed this feature, but are planning to re-introduce it into the next
Ganeti release. To allow for smooth up/downgrades, we will prepare for
the `-name` argument to carry additional parameters.

Signed-off-by: Rudolph Bott <r@bott.im>