[build] call cabal using a full path
`make distcheck' performs a so-called "VPATH build"[1], whereby it
copies everything in Makefile.am's DIRS in a subdirectory and performs
the build there, leaving the original source directory untouched.

Unfortunately by doing so it doesn't copy the Cabal files from the
source directory, and building subsequently fails.

Qualify the paths to Setup.lhs and ganeti.cabal using $(top_srcdir).
Also create a symlink for htest.hs, which was missing.

[1] https://www.gnu.org/software/automake/manual/html_node/VPATH-Builds.html

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>